Simplify disconnect duplication
